Avoid unnecessary renders on the Downloads section when an item is downloading
