+++ package.json
@@ -33,7 +33,7 @@
       "build/**/*",
       "node_modules/**/*"
     ],
-    "directories":{
+    "directories": {
       "buildResources": "assets"
     },
     "dmg": {
@@ -97,6 +97,7 @@
     "jest-environment-jsdom-fourteen": "0.1.0",
     "jest-resolve": "24.7.1",
     "jest-watch-typeahead": "0.3.0",
+    "memoize-one": "^5.0.5",
     "mini-css-extract-plugin": "0.5.0",
     "optimize-css-assets-webpack-plugin": "5.0.1",
     "pnp-webpack-plugin": "1.2.1",
@@ -113,6 +114,7 @@
     "react-jss": "^8.6.1",
     "react-redux": "^7.1.0",
     "react-visibility-sensor": "^5.1.1",
+    "react-window": "^1.8.5",
     "redux": "^4.0.1",
     "resolve": "1.10.0",
     "sass-loader": "7.1.0",
