+++ src/App/components/Downloads/DownloadedItems.js
@@ -1,30 +1,16 @@
 import React, { useContext, memo, useCallback } from 'react'
 import { HistoryContext } from '../../../Providers/HistoryProvider';
-import DownloadsItem from './DownloadsItem';
+import DownloadsItem from './Item';
 
 const DownloadedItems = ({ theme }) => {
   console.log('DownloadedItems updated')
   const { history } = useContext(HistoryContext);
     const items = [];
     for (let item in history) {
-      console.log(item, 'loop')
       const { id, date, name } = history[item];
-      items.push(<DownloadsItem id={id} date={date} name={name} theme={theme} status='downloaded' key={`downloaded${id}`} />);
+      items.push(<DownloadsItem id={id} date={date} name={name} theme={theme} status='downloaded' key={id} />);
     }
   items.sort((a, b) => b.props.date - a.props.date);
-  //console.log(history)
-  //const { history } = useContext(HistoryContext);
-  //console.log(history)
-  // const renderDownloads = useCallback(() => {
-  //   const items = [];
-  //   for (let item in history) {
-  //     const { id, date, name } = history[item];
-  //     items.push(<DownloadsItem id={id} date={date} name={name} theme={theme} status='downloaded' key={`downloaded${id}`} />);
-  //   }
-  //   //items.sort((a, b) => b.date - a.date);
-  //   console.log('sort')
-  //   return items.sort((a, b) => b.props.date - a.props.date);
-  // }, [])
   return (
     <div className='downloadMenu DownloadsItem'>
       {items}
@@ -32,25 +18,4 @@ const DownloadedItems = ({ theme }) => {
   );
 }
 
-const areEqual = (prevProps, nextProps) => {
-  console.log('equalllll ?')
-  // return false
-  const { history } = prevProps;
-  const { newHistory } = nextProps;
-  const isEqual = JSON.stringify(history) === JSON.stringify(newHistory)
-  console.log('==================')
-  console.log(isEqual)
-  console.log('==================')
-  return isEqual
-  // if (progress){
-  //   console.log('==================')
-  //   console.log(status === newStatus && progress.progress === newProgress.progress)
-  //   console.log('==================')
-  //   return status === newStatus && progress.progress === newProgress.progress
-  // }
-  // console.log('==================')
-  // console.log(status === newStatus)
-  // console.log('==================')
-  // return status === newStatus
-}
-export default memo(DownloadedItems, areEqual); 
+export default DownloadedItems;
