+++ src/App/components/Downloads/DownloadsInQueue.js
@@ -1,17 +1,19 @@
-import React, { useContext } from 'react';
-import DownloadsItem from './DownloadsItem';
+import React, { useContext, useCallback } from 'react';
+import DownloadsItem from './Item';
 import { DownloadQueueContext } from '../../../Providers/DownloadQueueProvider'
 
 const DownloadsInQueue = ({ theme }) => {
-  const { queue } = useContext(DownloadQueueContext);
+  const { queue, removeItemfromQueue } = useContext(DownloadQueueContext);
+  
   const renderDownloads = () => {
     if (queue.length === 0) return null
     return (
       <div className='downloadMenu DownloadsInQueue' style={{ marginBottom: '5vh' }}>
         {
           queue.map(item => {
-            const { id } = item;
-            return <DownloadsItem id={id} theme={theme} status='queued' key={`queued${id}`} />
+            const { id, fullTitle } = item;
+            const unQueue = () => removeItemfromQueue(id)
+            return <DownloadsItem id={id} name={fullTitle} theme={theme} cancel={unQueue} status='queued' key={id} />
           })
         }
       </div >
@@ -20,7 +22,7 @@ const DownloadsInQueue = ({ theme }) => {
   return (
     <React.Fragment>
       {renderDownloads()}
-    </React.Fragment>
+    </React.Fragment> 
   );
 }
